fix(history): list a merge only when it gave the record content of its own
A merge that only brought in a listed commit's content repeated that
commit and named another branch. Review also found no fixture with a
merge and no check of the commit date; both are now tested.

+++ internal/tui/view.go
@@ -435,7 +435,7 @@ func (m *Model) historyRows(w int) []string {
 		rows = append(rows, wrap("No commit here changed this file.", w)...)
 	}
 	for _, c := range read.commits {
-		entry(c.When.Format("2006-01-02 15:04"), c.Status, short(c.ID)[:min(len(c.ID), 7)]+"  "+c.Subject)
+		entry(c.When.Format("2006-01-02 15:04"), c.Status, c.ID[:min(len(c.ID), 7)]+"  "+c.Subject)
 	}
 	return rows
 }
